refactor wallet page
